Added typed viewports
Refractored viewpor manager to work with typed viewports
Fixed a LOT of viewports bugs

+++ src/draw/bfDrawObject2.cpp
@@ -174,7 +174,7 @@ BfDrawControlProxy::updateBuffer(
 void
 BfDrawControlProxy::draw(
     VkCommandBuffer combuffer,
-    size_t viewport_index,
+    uint32_t viewport_index,
     std::queue< std::shared_ptr< BfDrawObjectBase > >* q,
     size_t* offset,
     size_t* index_offset,
